secondary_backend/requirements.txt
openai
qdrant-client
python-dotenv
fastembed